p,r,t=map(float,input().split()) ci=p*((1+r/100)**t) print(format(ci,'.2f'))

+++ Compound_interest.py
@@ -1,3 +1,3 @@
-p,r,t=map(int,input().split())
+p,r,t=map(float,input().split())
 ci=p*((1+r/100)**t)
 print(format(ci,'.2f'))
